requirements.txt
dvc[azure]
requests
pandas
sqlalchemy
psycopg2-binary
joblib
numpy
certifi
scikit-learn
great_expectations
pyyaml
xgboost
imbalanced-learn
mlflow
torch
optuna
optuna-integration[mlflow]
uvicorn[standard]
pydantic
httpx
pytest
fastapi
pytest-asyncio